% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{bum_score}
\alias{bum_score}
\title{Scoring function generator (beta-uniform model)}
\usage{
bum_score(pvals, fdr = 0.01, threshold_pval, type = c("aggressive",
  "light"), bum_plot = FALSE, verbose = FALSE)
}
\arguments{
\item{pvals}{a numeric vector.Pp-values that will be used for learning
scoring function}

\item{fdr}{false discovery rate. It is used for calculating threshold p-value.
P-values less than threshold value will be scored positively.}

\item{threshold_pval}{custom p-value threshold}

\item{bum_plot}{whether or not plot BUM fitting plot}

\item{verbose}{be verbose}
}
\description{
Scoring function generator (beta-uniform model)
}
\seealso{
\code{\link{posterior_probabilities}}
}
