% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_tools.R
\name{estimate_network}
\alias{estimate_network}
\title{finds active module of given network}
\usage{
estimate_network(network, solver_time_factor = 1, ...)
}
\arguments{
\item{network}{igraph undirected graph with supplied gene names as vertex names}

\item{solver_time_factor}{solver will take time proportionally to this parameter}

\item{...}{other parameters of mwcsr::rmwcs}
}
\description{
finds active module of given network
}
