% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_tools.R
\name{permutation_freq}
\alias{permutation_freq}
\alias{BM_pvalue}
\title{computes frequencies of genes to be in a random network}
\usage{
permutation_freq(pvals, network, scoring_function = bum_score(pvals),
  permute_interactome = TRUE, permute_pvals = FALSE,
  n_permutations = 1000, permutation_method = c("igraph", "ppitools"),
  permutation_time_factor = 1, solver_time_factor = 1, plot = FALSE,
  simplify = TRUE, threads = parallel::detectCores())

BM_pvalue(pvals, network, scoring_function = BUM_score(pvals),
  threads = parallel::detectCores(), solver_time_factor = 1,
  simplify = TRUE, n_permutations = 100, permutation_method = c("igraph",
  "ppitools"), permutation_time_factor = 1, verbose = FALSE)
}
\arguments{
\item{pvals}{named vector of pvals or data frame containing columns \
gene, p_value and locus}

\item{network}{igraph undirected graph with supplied gene names as vertex names}

\item{permute_interactome}{should interactome be shuffled}

\item{permute_pvals}{should p-values be shuffled}

\item{n_permutations}{number of permutations to do}

\item{permutation_method}{igraph or ppitools, choose second choice if \
you want to change permutation_time_factor}

\item{permutation_time_factor}{permutator will take time proportionally to this parameter}

\item{solver_time_factor}{solver will take time proportionally to this parameter}

\item{plot}{whether to plot a module}

\item{simplify}{if FALSE sets p-value 0.5 to genes with unknown p-value}

\item{threads}{number of threads that could be used for computations}

\item{verbose}{verbosity}

\item{scf}{scoring function(e.g. bum_score output)}
}
\description{
computes frequencies of genes to be in a random network
}
