% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_tools.R
\name{plot_mwcs}
\alias{plot_mwcs}
\title{plots heat-mapped module}
\usage{
plot_mwcs(g, pps, log_space = TRUE, colors = 16)
}
\arguments{
\item{g}{igraph network}

\item{pps}{named vector of some posterior probabilies}

\item{log_space}{considering posteriors in log space}

\item{colors}{number of colors in heat map palette}
}
\description{
plots heat-mapped module
}
