% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_tools.R
\name{posterior_probabilities}
\alias{posterior_probabilities}
\title{computes p(U|BM)}
\usage{
posterior_probabilities(pvals, network, scoring_function = BUM_score(pvals),
  threads = parallel::detectCores(), solver_time_factor = 1,
  simplify = TRUE, verbose = FALSE, plot = TRUE)
}
\arguments{
\item{pvals}{named vector of pvals or data frame containing columns \
gene, p_value and locus}

\item{network}{igraph undirected graph with supplied gene names as vertex names}

\item{threads}{number of threads that could be used for computations}

\item{solver_time_factor}{solver will take time proportionally to this parameter}

\item{simplify}{if FALSE sets p-value 0.5 to genes with unknown p-value}

\item{verbose}{be verbose}

\item{plot}{whether to plot a module}
}
\description{
computes p(U|BM)
}
