% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_tools.R
\name{preprocess}
\alias{preprocess}
\title{preprocesses network and p-values}
\usage{
preprocess(pvals, network, simplify, scf)
}
\arguments{
\item{pvals}{named vector of pvals or data frame containing columns \
gene, p_value and locus}

\item{network}{igraph undirected graph with supplied gene names as vertex names}

\item{simplify}{if FALSE sets p-value 0.5 to genes with unknown p-value}

\item{scf}{scoring function(e.g. bum_score output)}
}
\description{
It is convinient to do before passing
network to estimate_network function, especially in case when set of genes
presetnted in pvals does not completely match to set of genes in network
}
