% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_tools.R
\name{posterior_probabilities}
\alias{posterior_probabilities}
\title{computes p(U|BM)}
\usage{
posterior_probabilities(pvals, network, scoring_function = BUM_score(pvals),
  threads = parallel::detectCores(), solver_time_factor = 1,
  simplify = TRUE, verbose = FALSE, plot = TRUE)
}
\arguments{
\item{pvals}{a named vector of p-values}

\item{network}{an igraph undirected graph object with vertex attribute "name"}

\item{threads}{calculations will be run in this number of threads}

\item{solver_time_factor}{solver of MWCS problem will work amount of time
proportionally to this factor}

\item{simplify}{if FALSE missing p-values will be filled by 0.5}

\item{verbose}{be verbose}
}
\description{
computes p(U|BM)
}
