% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_tools.R
\name{posterior_probabilities}
\alias{posterior_probabilities}
\title{computes posterior smth}
\usage{
posterior_probabilities(gene_pvals, network, permutations = 10000,
  scoring_function = BUM_score(gene_pvals),
  threads = parallel::detectCores(), verbose = FALSE, simplify = TRUE,
  solver_time_factor = 1, permutation_method = c("igraph", "package"),
  permutations_time_factor = 1)
}
\arguments{
\item{gene_pvals}{a named vector of p-values}

\item{network}{an igraph undirected graph object with vertex attribute "name"}

\item{threads}{calculations will be run in this number of threads}

\item{verbose}{be verbose}

\item{simplify}{if FALSE missing p-values will be filled by 0.5}

\item{solver_time_factor}{solver of MWCS problem will work amount of time
proportionally to this factor}

\item{permutation_method}{a method that will be used for permuations. 'package'
supports different time factors}

\item{permutation_time_factor}{time factor for permutations}
}
\description{
computes posterior smth
}
